upgrade packages and add 'use client'
