+++ apps/client/app/card/[id]/page.tsx
@@ -1,14 +1,20 @@
+'use client'
+
 import { MaybeCard } from '@maybe-finance/client/shared'
 import { type SharedType, superjson } from '@maybe-finance/shared'
 import { Button, LoadingSpinner, Tooltip } from '@maybe-finance/design-system'
 import Head from 'next/head'
 import { useState } from 'react'
 import { RiAnticlockwise2Line } from 'react-icons/ri'
-import env from '../../env'
+import env from '../../../env'
 import { NotFoundPage } from '@maybe-finance/client/features'
 
-export default function Card({ rawData }: { rawData: any }) {
+export default async function Card({ params: { id } }: { params: { id: string } }) {
     const [isFlipped, setIsFlipped] = useState(false)
+
+    const rawData = await fetch(`${env.NEXT_PUBLIC_API_URL}/v1/users/card/${id}`).then((data) =>
+        data.json()
+    )
 
     if (!rawData.data) return <NotFoundPage />
 
@@ -57,18 +63,6 @@ export default function Card({ rawData }: { rawData: any }) {
             </div>
         </>
     )
-}
-
-export async function getServerSideProps(context) {
-    const { id } = context.query
-
-    return {
-        props: {
-            rawData: await fetch(`${env.NEXT_PUBLIC_API_URL}/v1/users/card/${id}`).then((data) =>
-                data.json()
-            ),
-        },
-    }
 }
 
 Card.isPublic = true
