+++ apps/client/app/data-editor/page.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import type { ReactElement } from 'react'
 import { useMemo } from 'react'
 import {
