+++ apps/client/app/page.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import type { ReactElement } from 'react'
 import { useState } from 'react'
 import {
