+++ apps/client/app/plans/page.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import type { ReactElement } from 'react'
 import { useEffect } from 'react'
