+++ package.json
@@ -44,9 +44,7 @@
         "@polygon.io/client-js": "^6.0.6",
         "@popperjs/core": "^2.11.5",
         "@prisma/client": "^5.8.0",
-        "@sentry/node": "^7.22.0",
-        "@sentry/react": "^7.22.0",
-        "@sentry/tracing": "^7.22.0",
+        "@sentry/nextjs": "^7.95.0",
         "@storybook/core-server": "6.5.15",
         "@stripe/stripe-js": "^1.44.1",
         "@tanstack/react-query": "^4.19.1",
